Rename abc modules to network
Apparently, Python 2 doesn't like it when you give your (sub)modules the
same name as a standard library module. \shrug
