+++ neet/boolean/logicnetwork.py
@@ -11,7 +11,7 @@ Logic-based Networks
 import re
 from neet.python import long
 from neet.exceptions import FormatError
-from .abc import BooleanNetwork
+from .network import BooleanNetwork
 
 
 class LogicNetwork(BooleanNetwork):
